viz/ga: handle pivot category
Display pivot category in the Growth Advantage panel abstract and remove
the pivot category from x-axis of the plot since it is always 1 and does
not need to be plotted.¹

Currently assumes the last variant in `mlr.metadata.variants` is the
pivot category. Eventually, the evofr output will include an explicit
pivot category in the metadata and we can use the explicit value.

¹ https://bedfordlab.slack.com/archives/C036L7V615F/p1672875316627319?thread_ts=1672798375.781679&cid=C036L7V615F