Shift nCoV-specific settings to parameters
The aim is that there is nothing nCoV-specific in the library itself,
and all necessary information is provided by the config object introduced
here. There are still many hardcoded assumptions, but this removes
the most egregious. For instance, when there is a new variant in the
model data we need to update `variantColors` and `variantDisplayNames`
and this is now done by the parent component rather than having to
change the library itself.