Rename to evofr-viz + update README
The naming is still in flux, but evofr is much better than the previous effort! Note that the name defined in package.json is what matters for npm purposes, not the repo name.

The README was unchanged from the previous iteration of this code before we made it into a library.