Parse each model JSON independently
This makes the parsing logic simpler, especially as it relates to
dates. The test app (src/App.js) has been updated to show all the plots
supported by the current MLR / Renewal models.

Additionally `r_t` has been renamed `R` to both be more consistent
with the key used in the JSON and because the model is inferring big R
not little r!