Improve model loading + error display
Forecast JSONs typically load & parse in a few seconds so the main
thing to convey is if a model has failed to download, parse
or visualise, or if it's (still) loading. We accomplish the latter via
Auspice's spinning nextstrain-logo. The former is still tied to the
panel display component rather than the useModelData hook as (a) the
hook can't render to the DOM itself and (b) I didn't want an extra
Error component to have to be thought about when building these pages.
The upside of this approach is that where there should be a panel of
figures you now see an error message (or spinner); the downside being
that you may see duplicated errors if the error comes from the JSON
downloading / parsing rather than the visualisation itself.

Motivated by review comments by @joverlee521 in
https://github.com/nextstrain/forecasts-viz/pull/3