Tooltips (mk1)
This implements tooltips for frequencies graphs. They're useful as-is,
which is why I'm pushing up this commit, but this is not the final
version. They're quite hard to use at the moment (small capture area,
lots of elements sit over the capture element and swallow events), and
it would be better to have a tooltip whenever mousing over the graph
area which showed _all_ the frequency values for the given date.