Update-able graphs
Major overhaul of the approach we use to call D3 so that we can now
make partial updates to graphs. Currently the only implemented
partial update is the logit-toggle (frequencies graph).

This work ended up touching nearly all of the code base and included
a new preset-system for graph definitions, major additions to the
documentation, and consolidation of the d3 code itself.

Many decisions were made as part of this commit -- see NOTES.md for
more discussion.