Improve panel and legend sizes
This accomplishes (for now) the goal of reasonably nice styling
across a range of different datasets. The datasets we may receive
range in scale drastically: the number of locations (i.e. small
multiples) may be between 1 and >>20, the number of variants are between
a handful and >>100 etc. To have nice styling for every combination,
including different window sizes, is _hard_.