Listen for window resizes
This continues the design of the legend changing position but we now
use the outer div dimensions rather than the window dimensions; a much
better system as these graphs are designed to be reusable components.

This approach has some wins but also introduces some jank where
components update themselves ~200ms (or whatever debounce time we use)
later than expected. Using the observed height can be challenging
due to inherent race conditions in browser rendering - note that
the legend RHS display is currently buggy; I'll fix that in the
subsequent commit.