Remove RHS legend
I had been perusing a design where larger screens had the legend
rendered on the RHS. This had never quite worked as I had imagined,
and the preceding couple of commits had convinced me that it's no
longer worth the effort to try to make it work for the various use cases.
Perhaps we'll revisit it (or other legend designs) in the future.

In terms of rendering performance, the work here (and in the preceding
commits) causes each (d3) graph to be torn down & rebuilt after the
window dimensions change. This is debounced so that it only occurs once
with typical resizing behaviour.