Remove unnecessary re-rendering
This keeps the general approach in place, where the `useGraph` hook
decides what d3 functions to call based on what dependencies have
changed, but moves to a more robust deep-equality approach.
This has a couple of side-effects:

1. Any functions which are (within) dependencies must not be re-created
as they cannot be properly checked - e.g.:
```js
const a=()=>{};
isEqual(a(), a()) // false (they are not equal)
```

2.I  removed the use of `useMemo` as it was creating unnecessarily
complex code for very little gain. A better approach would be to
use a deep-equality memoization hook and wrap components in memo(), but
we can achieve almost the same result by comparing deps within
`useGraph`. The approach we use now results in more react component
updates, but no unnecessary d3 functions. The only change to the
re-rendering detailed below is the recalculations marked [*] would not
have happened if we kept the useMemo calls.

Introducing logging statements reveals the components re-rendered by
react (but not necessarily resulting in any DOM manipulation), when
running in production mode. This was done for two different scenarios:
(a) A single location for “General line graph (preset: frequency)”
(b) A single location for “Line graph using I_smooth” which uses graph
params defined in the calling parent.

```
<Panel> (incl. styling recalculation)
	Downloading & parsing model data JSON for Renewal
<Panel> (incl. styling recalculation)
	<Graph> (incl. params recalculation)
		useGraph
			Creating graph (for the first time)

*** Triggering <App> to re-render ***

<Panel> (incl. styling recalculation)
	<Graph> (incl. params recalculation)
		useGraph
			no d3 calls made at all

*** Logit toggle used (scenario (a) only) ***

<Panel> (incl. styling recalculation [*])
	<Graph> (incl. params recalculation [*])
		useGraph
			Updating graph as options have changed
```

Thanks to @joverlee521 for code review which motivated these changes.