Parse model update date
This serves two purposes: we expose it via modelData so that the calling
app can display this information, and it's used for the nowcasting -
forecasting crossover date.

Both pages in the test app are updated to surface this information.

This key was added to the (ncov) models via https://github.com/nextstrain/forecasts-ncov/pull/26