Add tests to cover date creation bug
The code for the creation of the dates array is unchanged, it's just
been refactored into a function which I've now written tests for.

The underlying bug is due to naive addition of days which results in
addition/subtraction of an hour as we cross daylight savings time
boundaries. Because the date object is at 00h00 this easily results in
duplicated / skipped days. For instance, the following are some date
ranges from the current implementation:

TZ: "Pacific/Auckland"
2023-09-23 -> 2023-09-25
  2023-09-23T00:00:00.000Z
  2023-09-23T23:00:00.000Z
  2023-09-24T23:00:00.000Z

2023-04-01 -> 2023-04-05
  2023-04-01T00:00:00.000Z
  2023-04-02T01:00:00.000Z
  2023-04-03T01:00:00.000Z
  2023-04-04T01:00:00.000Z

Thanks to @tsibley for identifying that this bug was related to daylight
savings changes.