Include toggles for both daily and weekly raw data
This expands on the previous commit to duplicate logic to toggles for both "daily_raw_freq_ and "weekly_raw_freq". The daily points are sized smaller and darker and the weekly points are sized larger and lighter. This seemed to work the best in terms of distinguishing both series, but the viz is really not optimized for this.