Legend hover highlights (existing) raw freq points
The chosen UI is to require one or both of the raw data toggles to be
on, and then hovering over a legend swatch "highlights" the corresponding
raw data.

If we only had one category of raw data (i.e. not daily _and_ weekly)
then i'd want to also show the dots on legend hover if the toggle was
off, but with two categories this isn't an obvious improvement.

The point sizing isn't quite right here. The next commit will address this.