Highlight lines & HPD when hovered on legend
The UI is an improvement in most cases, but for graphs where many
variants are at a consistently low level then this approach doesn't
always highlight the variant in a friendly way. A better approach would
be to modify the order of elements such that the focused variant is on
top, but that's a bigger change to the code.