Keep lineages as smaller multiples
Previously we had logic to make the frequency panels larger for lineages (ie when there are >20 variants). However, in practice this just results in a bunch of extraneous whitespace and makes it difficult to properly compare lineage frequencies across locations. Removing this bit of logic results in all frequency panels being the default size regardless of clade or lineage.