Allow sparse input data
The visualisations had assumed that the model inferences and forecasts
were done across a daily grid (i.e. each day had associated values),
and we leveraged this for the viz such that missing days would cause
breaks in the drawn frequency lines / CIs.

Recent model results for Influenza used a sparse grid and this assumption
caused empty frequency plots (because we only plotted frequency lines
for back-to-back days).

We solve this by deciding if the model data is (temporally) sparse or
not, and allow the JSON to define this if needed.