Improve ticks for sparse dates
The viz library was initially designed around nCoV model data which
would have values for every date in a certain range (i.e. no holes), and
so would display a label for the month (e.g. "Dec") at the x-value
corresponding to the first of the month.

We recently enabled display of sparse temporal data, and this approach
no longer works as it's rare for us to have values for the first of the
month. For sparse data we now display ticks for the first encountered
date of each month.

Note: sparse data may warrant exploration of temporal scales rather than
point scales, but that's left for another day. These point scales are
only going to be valid where the date (point) spacings are consistent!