User-definable parsing
Previously there was (limited) customisation of what data sites (keys) a
graph would use to display data, however the parsing of these sites was
not user-definable. We now allow the `useModelData` to define what sites
are parsed and how they're parsed.

The motivating case for these changes was <https://github.com/nextstrain/forecasts-flu/issues/13>
which wanted to use specific site (key) names in the JSON and also
relabel the associated toggles in the graph.

There are still hardcoded aspects and the API can be improved, but
since the overall direction of the project is up for grabs I'm
comfortable leaving these for another day.